transformer_layer refactor: moved the ffn block wrappers
